requirements.txt
cohere==5.13.6
pandas==2.2.3
streamlit==1.41.1
PyPDF2==3.0.1
pillow==11.1.0
python-dotenv==1.0.1
matplotlib==3.10.0
python-docx==1.1.2
openpyxl==3.1.5